# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: ExpertMatch API
  version: 1.0.0
  description: |
    GraphRAG-powered expert discovery and matching service for RFPs and team formation.
    This API enables automatic client generation for frontend applications using OpenAPI generators.
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: ""
    description: Current Server (auto-detected)
  - url: http://localhost:8093
    description: Local Development Server
  - url: http://192.168.0.73:8093
    description: Remote Development Server
  - url: https://api.expertmatch.com
    description: Production Server

# Security: Authentication and authorization are handled by Spring Gateway.
# Spring Gateway validates JWT tokens and populates user information in HTTP headers:
# - X-User-Id: User identifier (required)
# - X-User-Roles: Comma-separated list of roles (optional, e.g., "ROLE_USER,ROLE_ADMIN")
# - X-User-Email: User email address (optional)
...
components:
  parameters:
    X-User-Id:
      name: X-User-Id
      in: header
      required: false
      description: User identifier populated by Spring Gateway after authentication. If not provided, anonymous user will be used.
      schema:
        type: string
        default: "anonymous-user"
        example: "user-123"
    X-User-Roles:
      name: X-User-Roles
      in: header
      required: false
      description: Comma-separated list of user roles populated by Spring Gateway
      schema:
        type: string
        example: "ROLE_USER,ROLE_ADMIN"
    X-User-Email:
      name: X-User-Email
      in: header
      required: false
      description: User email address populated by Spring Gateway
      schema:
        type: string
        example: "user@example.com"
  schemas:
...
    Chat:
      type: object
      required:
        - id
        - userId
        - name
        - isDefault
        - createdAt
        - updatedAt
        - lastActivityAt
        - messageCount
      properties:
        id:
          type: string
          description: Chat ID (24-character hexadecimal string)
          example: "507f1f77bcf86cd799439011"
        userId:
          type: string
          description: User ID who owns the chat
        name:
          type: string
          description: Chat name
          example: "My Chat"
        isDefault:
          type: boolean
          description: Whether this is the user's default chat
        createdAt:
          type: string
          format: date-time
          description: Chat creation timestamp (ISO 8601)
        updatedAt:
          type: string
          format: date-time
          description: Chat last update timestamp (ISO 8601)
        lastActivityAt:
          type: string
          format: date-time
          description: Last activity timestamp (ISO 8601)
        messageCount:
          type: integer
          description: Number of messages in the chat
    CreateChatRequest:
      type: object
      properties:
        name:
          type: string
          description: Optional chat name. If not provided, a default name will be generated.
          example: "My New Chat"
    UpdateChatRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: New chat name
          example: "Updated Chat Name"
    ChatListResponse:
      type: object
      required:
        - chats
        - total
      properties:
        chats:
          type: array
          description: List of chats
          items:
            $ref: '#/components/schemas/Chat'
        total:
          type: integer
          description: Total number of chats
    ConversationMessage:
      type: object
      required:
        - id
        - chatId
        - messageType
        - role
        - content
        - sequenceNumber
        - createdAt
      properties:
        id:
          type: string
          description: Message ID
        chatId:
          type: string
          description: Chat ID this message belongs to
        messageType:
          type: string
          enum: [ user, assistant, system ]
          description: Message type
        role:
          type: string
          description: Message role (user, assistant, system)
        content:
          type: string
          description: Message content
        sequenceNumber:
          type: integer
          description: Message sequence number in the conversation
        tokensUsed:
          type: integer
          nullable: true
          description: Number of tokens used (null for user messages)
        createdAt:
          type: string
          format: date-time
          description: Message creation timestamp (ISO 8601)
    PageInfo:
      type: object
      required:
        - number
        - size
        - totalElements
        - totalPages
      properties:
        number:
          type: integer
          description: Current page number (0-indexed)
        size:
          type: integer
          description: Page size
        totalElements:
          type: integer
          description: Total number of elements
        totalPages:
          type: integer
          description: Total number of pages
    ConversationHistoryResponse:
      type: object
      required:
        - chatId
        - messages
        - page
      properties:
        chatId:
          type: string
          description: Chat ID
        messages:
          type: array
          description: List of conversation messages
          items:
            $ref: '#/components/schemas/ConversationMessage'
        page:
          $ref: '#/components/schemas/PageInfo'
    DeleteChatResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: Whether the deletion was successful
        message:
          type: string
          description: Response message

    # Ingestion-related schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        status:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp
        path:
          type: string
          description: API path that caused the error
        details:
          type: object
          description: Additional error details
          additionalProperties: true
paths:
...
  /chats:
    post:
      tags:
        - Chat Management
      summary: Create a new chat
      description: |
        Creates a new chat for the authenticated user.
        If name is not provided, a default name will be generated.
      operationId: createChat
      parameters:
        - $ref: '#/components/parameters/X-User-Id'
        - $ref: '#/components/parameters/X-User-Roles'
        - $ref: '#/components/parameters/X-User-Email'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatRequest'
      responses:
        '201':
          description: Chat created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Chat Management
      summary: List all chats
      description: Returns all chats for the authenticated user, including the default chat.
      operationId: listChats
      parameters:
        - $ref: '#/components/parameters/X-User-Id'
        - $ref: '#/components/parameters/X-User-Roles'
        - $ref: '#/components/parameters/X-User-Email'
      responses:
        '200':
          description: List of chats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /chats/{chatId}:
    get:
      tags:
        - Chat Management
      summary: Get chat details
      description: |
        Returns details for a specific chat.
        Verifies that the chat belongs to the authenticated user.
      operationId: getChat
      parameters:
        - $ref: '#/components/parameters/X-User-Id'
        - $ref: '#/components/parameters/X-User-Roles'
        - $ref: '#/components/parameters/X-User-Email'
        - name: chatId
          in: path
          required: true
          description: 24-character hexadecimal chat ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Chat details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          description: Invalid chat ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied to chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Chat Management
      summary: Update chat
      description: |
        Updates a chat, typically to rename it.
        Verifies that the chat belongs to the authenticated user.
      operationId: updateChat
      parameters:
        - $ref: '#/components/parameters/X-User-Id'
        - $ref: '#/components/parameters/X-User-Roles'
        - $ref: '#/components/parameters/X-User-Email'
        - name: chatId
          in: path
          required: true
          description: 24-character hexadecimal chat ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: "507f1f77bcf86cd799439011"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChatRequest'
      responses:
        '200':
          description: Chat updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          description: Invalid chat ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied to chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Chat Management
      summary: Delete a chat
      description: |
        Deletes a chat. Cannot delete the default chat.
        Verifies that the chat belongs to the authenticated user.
      operationId: deleteChat
      parameters:
        - $ref: '#/components/parameters/X-User-Id'
        - $ref: '#/components/parameters/X-User-Roles'
        - $ref: '#/components/parameters/X-User-Email'
        - name: chatId
          in: path
          required: true
          description: 24-character hexadecimal chat ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Chat deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteChatResponse'
        '400':
          description: Invalid chat ID format or attempting to delete default chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied to chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /chats/{chatId}/history:
    get:
      tags:
        - Chat Management
      summary: Get conversation history
      description: |
        Returns paginated conversation history for a chat.
        Verifies that the chat belongs to the authenticated user.
      operationId: getHistory
      parameters:
        - $ref: '#/components/parameters/X-User-Id'
        - $ref: '#/components/parameters/X-User-Roles'
        - $ref: '#/components/parameters/X-User-Email'
        - name: chatId
          in: path
          required: true
          description: 24-character hexadecimal chat ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: "507f1f77bcf86cd799439011"
        - name: page
          in: query
          required: false
          description: Page number (0-indexed)
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
        - name: size
          in: query
          required: false
          description: Page size (1-100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: sort
          in: query
          required: false
          description: Sort criteria (e.g., 'sequence_number,asc')
          schema:
            type: string
            default: "sequence_number,asc"
          example: "sequence_number,asc"
      responses:
        '200':
          description: Conversation history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationHistoryResponse'
        '400':
          description: Invalid chat ID format or pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied to chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...